Change: コンポーネントを分割
